requirements.txt
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
cloudinary==1.44.1
colorama==0.4.6
Deprecated==1.2.18
Flask==3.1.1
Flask-Limiter==3.12
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==25.0
pillow==11.3.0
Pygments==2.19.2
requests==2.32.4
rich==13.9.4
six==1.17.0
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.17.2
gunicorn==21.2.0
pytz==2025.2
